requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.2
certifi==2020.6.20
colorama==0.4.3
Django==3.1.1
django-bootstrap4==2.2.0
django-filter==2.4.0
djangorestframework==3.12.1
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
Pillow==7.2.0
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
wincertstore==0.2
wrapt==1.11.2
